plotting/power.py: Silence a FutureWarning
FutureWarning: The default dtype for empty Series will be 'object'
instead of 'float64' in a future version. Specify a dtype explicitly to
silence this warning.

+++ plotting/power.py
@@ -94,9 +94,9 @@ def plot(num_rows=0, row_pos=1):
             pass
 
         try:
-            df_power['duration'] = pd.Series()
-            df_power['energy_diff'] = pd.Series()
-            df_power['power'] = pd.Series()
+            df_power['duration'] = pd.Series(dtype='float64')
+            df_power['energy_diff'] = pd.Series(dtype='float64')
+            df_power['power'] = pd.Series(dtype='float64')
 
             color = ['r', 'y', 'b']
             i = 0
